+++ sdk-go/dephealth/checks/factories_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/alicebob/miniredis/v2"
 
-	"github.com/BigKAA/topologymetrics/dephealth"
+	"github.com/BigKAA/topologymetrics/sdk-go/dephealth"
 )
 
 func TestPostgresFactory_URLPassedAsDSN(t *testing.T) {