+++ sdk-go/dephealth/checks/grpc_test.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc"
 	healthpb "google.golang.org/grpc/health/grpc_health_v1"
 
-	"github.com/BigKAA/topologymetrics/dephealth"
+	"github.com/BigKAA/topologymetrics/sdk-go/dephealth"
 )
 
 // testHealthServer implements the gRPC Health service for testing.