+++ sdk-go/dephealth/checks/kafka_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/BigKAA/topologymetrics/dephealth"
+	"github.com/BigKAA/topologymetrics/sdk-go/dephealth"
 )
 
 func TestKafkaChecker_Check_ConnectionRefused(t *testing.T) {