+++ sdk-go/dephealth/checks/mysql.go
@@ -8,7 +8,7 @@ import (
 
 	_ "github.com/go-sql-driver/mysql" // MySQL driver
 
-	"github.com/BigKAA/topologymetrics/dephealth"
+	"github.com/BigKAA/topologymetrics/sdk-go/dephealth"
 )
 
 // MySQLOption configures the MySQLChecker.