+++ sdk-go/dephealth/checks/mysql_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/DATA-DOG/go-sqlmock"
 
-	"github.com/BigKAA/topologymetrics/dephealth"
+	"github.com/BigKAA/topologymetrics/sdk-go/dephealth"
 )
 
 func TestMySQLChecker_Check_PoolMode(t *testing.T) {