+++ sdk-go/dephealth/contrib/redispool/redispool.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/redis/go-redis/v9"
 
-	"github.com/BigKAA/topologymetrics/dephealth"
-	"github.com/BigKAA/topologymetrics/dephealth/checks"
+	"github.com/BigKAA/topologymetrics/sdk-go/dephealth"
+	"github.com/BigKAA/topologymetrics/sdk-go/dephealth/checks"
 )
 
 // FromClient создаёт Option для мониторинга Redis через существующий *redis.Client.